Create Domain checking
Domain checking
You are planning your summer vacation and need to secure some permits. So you want to make sure that a site is from the US government to be sure of its legitimacy. You've been told they all use the ".gov" domain, and that only government sites can use it.

Instructions

Check that the given url contains ".gov" as a substring. Use the variable toFind to do that.