Create Security checking
Security checking
You want to make sure that the site you are trying to visit is secure. All websites which try to be secure start with https://. Those that don't start with http:// (notice the absence of the s).

Instructions

Assign the correct index of the character 's' in the url string to the variable charPosition.
Use the appropriate String method on the url string to extract the character at that position and print it.