Create Squaring
Squaring
Working part-time 5 hours a day, 5 days a week, you often have to remember the square of 5. Rather than putting it everywhere in our code, let's develop a method to make it easier.

Instructions

Give the proper type to return for the custom method.
Call that method, saving the result as hours.