Create Power 9000
Power 9000
You are attempting to read the power of a multimeter in a factory. You have to make sure that everyone knows when it gets above 9000. Otherwise, nothing should happen.

Instructions

Enter the keyword allowing to check if a condition is true.
Use the correct operator to only display a word if it is greater than 9000.