Create Grading
Grading
In your school, students get their grades online. You want to make sure that students get complimented when strictly above 11, but also that they get a clear message when it is failed. The grades go from 0 to 20, and 10 is the passing grade.

Instructions

Use the correct operator and number to make sure that grades strictly above 11 get a compliment.
Use the correct keywords to make sure that it is entered only if the previous condition is false and something else is met.
Use the proper keyword to catch all other cases.