Create Weather Reporter
Weather Reporter
You've found a way to retrieve the weather report for your location. You want to have a small program that tells you what item to use in different types of weather.

Instructions
Use the correct keyword to make it a case to check.
Put in the right constant to print only when it is "Sunny".
Write the correct keyword down to make sure only one println() will take effect.