Create Spelling bee
Spelling bee
You have entered a spelling competition! To help you, you want to make a program which will decompose the words into letters.

Instructions

Apply the correct method to retrieve the length of toSpell.
Use the correct keyword to start a for loop.
Set the correct starting value for i - you want to start from the first character.
Write the correct method to get the character at the ith position of toSpell.