Create Batch Checking
Batch Checking
You are responsible for the quality checking of different batches of chocolate boxes. You are aware that one is too heavy, but you don't know which one. Print them out in order to find out which one is above 220g.

Instructions

Use the correct symbol to mark the difference between the elements and the collection.
Call the proper variable to print how much each batch weighs.