Create Odd summing
Odd summing
You wish to know how quickly the sum of the odd number grows up to 100. Why not write a code that does that?

Instructions

Enter the correct keyword to create the loop.
Call the appropriate variable to store the total reached.
Use the fitting operation to have i increase by 2 each run.