Create Reversing!
Reversing!
You want to perform a practical joke on your website and replace words with their inverse. You are testing your method before deploying.

Instructions

Give the correct name for the method.
Enter the appropriate parameter type for the method.
Choose the correct starting value for i in the loop.