Create New price
New price
Your government just introduced new taxes on your goods. You want to update all of the prices in your inventory at one time, adding half of its value to it.

Instructions

Give the correct type to your function.
Return one and a half multiple of the argument.
Use the appropriate type of for loop.